+++ TidalPy/utilities/performance/array/__init__.py
@@ -1,2 +1,2 @@
 # Import `interp` function
-from TidalPy.utilities.performance.array.interp import interp
+from TidalPy.utilities.performance.array.interp import interp, interp_complex